LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/abort-while-navigating.window-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8800/common/blank.html due to access control checks.
CONSOLE MESSAGE: Fetch API cannot load http://localhost:8800/common/blank.html due to access control checks.

Harness Error (TIMEOUT), message = null

PASS document.open() aborts documents that are navigating through Location (XMLHttpRequest)
PASS document.open() aborts documents that are navigating through Location (fetch())
PASS document.open() aborts documents that are navigating through Location (image loading)
TIMEOUT document.open() aborts documents that are navigating through iframe loading (XMLHttpRequest) Test timed out
PASS document.open() aborts documents that are navigating through iframe loading (fetch())
PASS document.open() aborts documents that are navigating through iframe loading (image loading)
PASS document.open() aborts documents that are queued for navigation through .click() (XMLHttpRequest)
PASS document.open() aborts documents that are queued for navigation through .click() (fetch())
PASS document.open() aborts documents that are queued for navigation through .click() (image loading)
